[dagster-airlift] Move snippets to snippets folder
