+++ examples/docs_snippets/docs_snippets/integrations/airlift/tutorial/all_proxied.yaml
