+++ examples/docs_snippets/docs_snippets/integrations/airlift/tutorial/custom_operator_examples/proxied_state/custom_proxy_example.yaml
