+++ examples/docs_snippets/docs_snippets/integrations/airlift/tutorial/dbt_proxied.yaml
