+++ examples/docs_snippets/docs_snippets/integrations/airlift/tutorial/rebuild_customers_list.yaml
